add shift components to shift route
