+++ frontend/src/App.css
@@ -3,9 +3,9 @@
 }
 
 .App-header {
-  /* align-items: center; */
-  /* flex-direction: column; */
-  /* justify-content: center; */
+  align-items: center; 
+  flex-direction: column;
+  justify-content: center;
 }
 
 nav {
